backend/requirements.txt
Flask
flask-cors
requests
python-dotenv
loguru
openai==0.27.10